requirements.txt
streamlit==1.40.0
pandas==2.2.3
numpy==1.26.4
faiss-cpu==1.10.0
sentence-transformers==3.4.1
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
joblib==1.4.2
requests==2.32.3
urllib3==2.2.3
aiohttp==3.11.11
httpx==0.28.1
pydantic==2.10.6
pydantic-core==2.27.2
SQLAlchemy==2.0.37
python-dateutil==2.9.0.post0
openai-whisper==20240930
scipy==1.15.1
sounddevice==0.5.1
deep-translator==1.11.4
langdetect==1.0.9
peft

#you have to install ffmpeg either using choclatey or downloading from the original website.